+++ WikipediaUnitTests/XCTestCase+WMFBundleConvenience.swift
@@ -1,0 +1,8 @@
+
+import Foundation
+
+extension XCTest {
+    func swiftWmf_bundle() -> Bundle {
+        return Bundle(for: XCTest.self)
+    }
+}
